Add tests to ensure that the error handling works. This commit contains pieces of code that are for debbuging, the next commit will remove then.
It's possible to test the framing error by executing error_gen.py. This script will send a picture to the ARM and after 100 bytes, two randon numbers with parity, bytesize and baudrate incorrects.
If the frame error occurs, the yellow light will blink.

To test the overrun, is necessary only to press the blue button. This will call a interruption that will do nothing for 2 ms.
If the overrun error occurs, the yellow light will blink.

The next commit will remove all the tests from the repository.